Invasion Event data

A tibble containing the introduction events present in the Shiny app. This dataset can be useful when working with the database programmatically.

Usage

inv_tbl

Format

inv_tbl

A tibble data frame with 4,834 rows and 18 columns:

InvID

Unique Identifier of an introduction event numeric

SpeciesID

Identifier linking a species to a valid taxonomic

Year

Year of the first repported occurence

Country

Country where the invasion event occurred

Status

Origin status of the species (i.e. Non-indigenous or Cryptogenic)

Release,Escape, Transport_contaminant, Transport_stowaway, Corridor, Unaided, Unknown

Pathways and vectors of introduction

MSFD_region

Name of the MSFD region. See here for details

MSFD_subregion

Name of the MSFD region. See here for details

Ecoregion_Code

Code of the Ecoregion (sensus Marine Ecoregion of the World according to Spalding et al. 2007) of the corresponding MSFD sub-region.

Source

Reference indicating that the species is a cryptogenic or non-indigenous species in the area

DB

Reference (e.g., database, review, or synthesis publication) compiling or citing the Source that indicates the species is cryptogenic or non-indigenous in the area

Source

E-MARINADE package